Move GridRenderer types into new file
