+++ packages/grid/src/GridRenderer.ts
@@ -4,87 +4,16 @@ import memoizeClear from './memoizeClear';
 import GridUtils, { Token, TokenBox } from './GridUtils';
 import GridColorUtils from './GridColorUtils';
 import { isExpandableGridModel } from './ExpandableGridModel';
-import {
-  GridColor,
-  GridColorWay,
-  GridTheme,
-  NullableGridColor,
-} from './GridTheme';
-import GridModel from './GridModel';
-import GridRange from './GridRange';
-import GridMetrics, {
-  BoxCoordinates,
-  Coordinate,
-  VisibleIndex,
-} from './GridMetrics';
+import { GridColor, GridColorWay, NullableGridColor } from './GridTheme';
+import { BoxCoordinates, Coordinate, VisibleIndex } from './GridMetrics';
 import { isEditableGridModel } from './EditableGridModel';
-import type { GridSeparator } from './mouse-handlers/GridSeparatorMouseHandler';
-import { DraggingColumn } from './mouse-handlers/GridColumnMoveMouseHandler';
 import GridColumnSeparatorMouseHandler from './mouse-handlers/GridColumnSeparatorMouseHandler';
 import { BoundedAxisRange } from './GridAxisRange';
-
-export type EditingCellTextSelectionRange = [start: number, end: number];
+import { GridRenderState } from './GridRendererTypes';
 
 type NoneNullColumnRange = { startColumn: number; endColumn: number };
 
 type NoneNullRowRange = { startRow: number; endRow: number };
-
-export type EditingCell = {
-  // Index of the editing cell
-  column: VisibleIndex;
-  row: VisibleIndex;
-
-  // Selection within the text
-  selectionRange?: EditingCellTextSelectionRange;
-
-  // The value to use for the edit
-  value: string;
-
-  // Whether the selection was triggered with a quick edit action (e.g. Start typing with the cell in focus)
-  isQuickEdit?: boolean;
-};
-
-export type GridRenderState = {
-  // Width and height of the total canvas area
-  width: number;
-  height: number;
-
-  // The canvas context
-  context: CanvasRenderingContext2D;
-
-  // The grid theme
-  theme: GridTheme;
-
-  // The model used by the grid
-  model: GridModel;
-
-  // The grid metrics
-  metrics: GridMetrics;
-
-  // Location of the mouse on the grid
-  mouseX: Coordinate | null;
-  mouseY: Coordinate | null;
-
-  // Where the keyboard cursor is located
-  cursorColumn: VisibleIndex | null;
-  cursorRow: VisibleIndex | null;
-
-  // Currently selected ranges
-  selectedRanges: readonly GridRange[];
-
-  // Currently dragged column/row information
-  draggingColumn: DraggingColumn | null;
-  draggingColumnSeparator: GridSeparator | null;
-  draggingRow: VisibleIndex | null;
-  draggingRowOffset: number | null;
-  draggingRowSeparator: GridSeparator | null;
-
-  // The currently editing cell
-  editingCell: EditingCell | null;
-  isDraggingHorizontalScrollBar: boolean;
-  isDraggingVerticalScrollBar: boolean;
-  isDragging: boolean;
-};
 
 /* eslint react/destructuring-assignment: "off" */
 /* eslint class-methods-use-this: "off" */
